server mod
git-svn-id: http://svn.kcn.cn/repos/kbs/trunk@10181 de4889a5-8ad4-4bf3-bd87-57997c21d515

+++ kbs_bbs/innbbsd/receive_article.c
@@ -266,7 +266,7 @@ int receive_article()
         innbbsdlog(" :Err: your bbsid %s too long\n", MYBBSID);
         return 0;
     }
-    sprintf(xpath, "%s!%.*s", MYBBSID, sizeof(xpath) - strlen(MYBBSID) - 2, PATH);
+    sprintf(xpath, "%s!%.*s", MYBBSID, (int)(sizeof(xpath) - strlen(MYBBSID) - 2), PATH);
     PATH = xpath;
     for (pathptr = PATH; pathptr != NULL && (pathptr = strstr(pathptr, ".edu.tw")) != NULL;) {
         if (pathptr != NULL) {
